Refactor runtime episode contracts
